Corporate Design: knitr Chunk Options

Set knitr chunk options in accordance with corporate design.

Usage

cd_knit_chunk_opts(
  twocolumn = default(rmarkdown::metadata$twocolumn, FALSE),
  fig.width = NULL,
  fig.height = NULL,
  fig.pos = "h",
  message = FALSE,
  echo = FALSE
)

Arguments

twocolumn

Use two-column mode? This will affect fig.width and fig.height.

fig.width

Figure width. If NULL, automatically chosen based on twocolumn.

fig.height

Figure height If NULL, automatically chosen based on twocolumn.

fig.pos

Floating modifier for figures

message

Should messages be shown?

echo

Should echo be shown?

Value

This function is called for its side effects and returns NULL, invisibly.

Examples

cd_knit_chunk_opts()